+++ backend/django_project/settings.py
@@ -40,6 +40,11 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
+    # 3rd party apps
+    "rest_framework",
+    "rest_framework.authtoken",
+    "corsheaders",
+    "djoser",
     # Local apps
     "users.apps.UsersConfig",
     "bookmarks.apps.BookmarksConfig",