+++ changelog.md
@@ -2,6 +2,7 @@
 
 - 12:50 - Created `users` app with `CustomUser` model.
 - 14:00 - Created `bookmarks` app with `Folder` and `Bookmark` models.
+- 19:00 - Built DRF API endpoints `bookmarks` and `bookmarks/create_from_telegram`.
 
 ### 03.04.2023, Mo